create core package and move all relevant files there, optimise imports
