+++ iris/api/core/screen.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from helpers.image_search import SCREEN_WIDTH, SCREEN_HEIGHT
+from iris.api.helpers.image_search import SCREEN_WIDTH, SCREEN_HEIGHT
 from region import Region, get_region
